Fix pull request todos
