+++ src/app/constants/colors.js
@@ -1,0 +1,6 @@
+const red = '#E3342F';
+const green = '#38C172';
+const purple = '#9561E2';
+const white = '#FFF';
+
+export { red, green, purple, white };
